Rename `CardMedia/Image`
