+++ packages/react-microlink/src/components/Card/CardMedia/index.js
@@ -2,7 +2,7 @@ import React, {Component, Fragment} from 'react'
 
 import {getUrlPath} from '../../../utils'
 
-import Image from './image'
+import Image from './Image'
 import Video from './Video'
 import {ImageLoadCatcher} from './loader'
